+++ src/menu.rs
@@ -160,8 +160,8 @@ fn create_logo_scene(window_size: Size) -> Scene<Texture> {
     );
     let mut sprite = Sprite::from_texture(Rc::clone(&tex));
     sprite.set_position(
-        f64::from(window_size.width) / 2.0,
-        f64::from(window_size.height) / 2.0 - 120.0,
+        window_size.width / 2.0,
+        window_size.height / 2.0 - 120.0,
     );
     sprite.set_scale(0.4, 0.4);
     sprite.set_opacity(0.0);
@@ -197,7 +197,7 @@ pub fn run(mut window: &mut PistonWindow, mut opengl: &mut GlGraphics, window_si
         music::play_music(&Music::Menu, music::Repeat::Forever);
 
         let mut menu_selection = MenuSelection::Play;
-        let menu_align = f64::from(window_size.width) / 2.0 - 120.0;
+        let menu_align = window_size.width / 2.0 - 120.0;
         while let Some(event) = window.next() {
             if let Some(args) = event.render_args() {
                 opengl.draw(args.viewport(), |context, graphics| {
